requirements.txt
requests==2.32.3
snowflake-connector-python==3.18.0
python-dotenv==1.1.0
